fix: change to type guard in fetchUserInfo
