+++ app/ui/user/user-info-card/init-user-announce.tsx
@@ -1,0 +1,13 @@
+function InitUserAnnounce() {
+  return (
+    <div className="h-44 text-gray-6 flex items-center justify-center">
+      <p className="text-center">
+        성적표 PDF 업로드를 통해
+        <br />
+        졸업사정 결과를 확인할 수 있어요.
+      </p>
+    </div>
+  );
+}
+
+export default InitUserAnnounce;
