+++ app/ui/user/user-info-card/user-info-card.tsx
@@ -1,74 +1,20 @@
-import { getPercentage } from '@/app/utils/chart.util';
-import PieChart from '../../view/molecule/pie-chart/pie-chart';
 import { fetchUserInfo } from '@/app/business/user/user.query';
-import { MAJOR_NOTATION } from '@/app/utils/key/result-category.key';
+import { InitUserInfoResponse, UserInfoResponse } from '@/app/business/user/user.type';
+import InitUserAnnounce from './init-user-announce';
+import UserInfoContent from './user-info-content';
 
 async function UserInfoCard() {
-  const {
-    studentNumber,
-    studentName,
-    completionDivision: majors,
-    totalCredit,
-    takenCredit,
-    graduated,
-  } = await fetchUserInfo();
+  const data = await fetchUserInfo();
 
-  const displaySeveralMajor = (notation: 'major' | 'title'): React.ReactNode => {
-    if (majors) {
-      return majors.map((major, index) => {
-        const { major: majorName, majorType } = major;
+  function isInitUser(x: UserInfoResponse | InitUserInfoResponse): x is InitUserInfoResponse {
+    return typeof x.studentName === null;
+  }
 
-        return <li key={index}>{notation === 'title' ? MAJOR_NOTATION[majorType] : majorName}</li>;
-      });
-    }
-  };
+  function renderUserInfo(data: UserInfoResponse | InitUserInfoResponse) {
+    isInitUser(data) ? <InitUserAnnounce /> : <UserInfoContent data={data} />;
+  }
 
-  return (
-    <>
-      {studentName ? (
-        <>
-          <p className="font-bold text-sm md:text-xl">
-            졸업필요학점보다 <span className="text-point-blue">{totalCredit - takenCredit}</span>학점이 부족합니다.
-          </p>
-          <div className="flex border-t-2 my-4 py-4 justify-between items-center">
-            <div className="flex font-medium text-xs md:text-lg gap-4 md:gap-14 ">
-              <ul className="text-gray-6 flex flex-col gap-1">
-                <li>이름</li>
-                <li>학번</li>
-                {displaySeveralMajor('title')}
-                <li>졸업필요학점</li>
-                <li>총 이수 학점</li>
-                <li>졸업가능여부</li>
-              </ul>
-              <ul className="flex flex-col gap-1">
-                <li>{studentNumber}</li>
-                <li>{studentName}</li>
-                {displaySeveralMajor('major')}
-                <li>{totalCredit}</li>
-                <li>{takenCredit}</li>
-                <li>{graduated ? '가능' : '불가능'}</li>
-              </ul>
-            </div>
-            <div className="md:px-10">
-              <PieChart percentage={getPercentage(takenCredit, totalCredit)} />
-            </div>
-          </div>
-          <p className="text-gray-6 text-xs">
-            * 서비스의 결과는 공식적인 효력을 갖지 않습니다. 정확한 졸업사정결과는 소속 단과대 교학팀에서의 확인을
-            권장합니다.
-          </p>
-        </>
-      ) : (
-        <div className="h-44 text-gray-6 flex items-center justify-center">
-          <p className="text-center">
-            성적표 PDF 업로드를 통해
-            <br />
-            졸업사정 결과를 확인할 수 있어요.
-          </p>
-        </div>
-      )}
-    </>
-  );
+  return <>{renderUserInfo(data)}</>;
 }
 
 export default UserInfoCard;
